refactor shared code filenames as first step to improved organisation
