+++ classic_mac/discovery.c
@@ -1,8 +1,8 @@
 #include "discovery.h"
 #include "logging.h"
 #include "protocol.h"
-#include "../shared/discovery_logic.h"
-#include "peer_mac.h"
+#include "../shared/discovery.h"
+#include "peer.h"
 #include "network.h"
 #include "dialog.h"
 #include "dialog_peerlist.h"
