+++ classic_mac/peer.c
@@ -1,4 +1,4 @@
-#include "peer_mac.h"
+#include "peer.h"
 #include "logging.h"
 #include <stddef.h>
 #include <stdbool.h>
