--- posix/logging.h
@@ -1,5 +1,0 @@
-#ifndef LOGGING_H
-#define LOGGING_H 
-void log_message(const char *format, ...);
-void display_user_message(const char *format, ...);
-#endif
