+++ shared/discovery.h
