+++ shared/logging.c
@@ -1,4 +1,4 @@
-#include "logging_shared.h"
+#include "logging.h"
 static Boolean g_show_debug_output = false;
 void set_debug_output_enabled(Boolean enabled) {
     g_show_debug_output = enabled;
