+++ shared/messaging.c
@@ -1,4 +1,4 @@
-#include "messaging_logic.h"
+#include "messaging.h"
 #include "protocol.h"
 #include "logging.h"
 #include <string.h>
