+++ shared/messaging.h
