+++ shared/peer.h
